requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
class_resolver==0.5.4
click==8.1.8
click-default-group==1.2.4
cloudpickle==3.1.1
colorlog==6.9.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
databricks-sdk==0.44.1
dataclasses-json==0.6.7
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
docdata==0.0.4
docker==7.1.0
executing==2.2.0
filelock==3.17.0
Flask==3.1.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
gunicorn==23.0.0
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
matplotlib-inline==0.1.7
mlflow==2.20.3
mlflow-skinny==2.20.3
more-click==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.2
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
optuna==4.2.0
owlready2==0.47
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pykeen==1.11.0
pyparsing==3.2.1
pystow==0.7.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
rdflib==7.1.3
requests==2.32.3
rsa==4.9
scikit-learn==1.6.1
scikit-posthocs==0.11.4
scipy==1.15.1
seaborn==0.13.2
setuptools==75.8.0
six==1.17.0
sklearn-compat==0.1.3
smmap==5.0.2
SQLAlchemy==2.0.37
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
torch==2.5.1
torch-geometric==2.6.1
torch-max-mem==0.1.3
torch-ppr==0.0.8
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wittgenstein==0.3.4
wrapt==1.17.2
yarl==1.18.3
zipp==3.21.0
